+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change logs
 
+## 2023.02.04 v1.1.0
+
+1. export `parseYml` & `stringifyYml`
+2. throw out errors
+3. fix dependencies
+
 ## 2023.02.03 v1.0.0
 
 1. create project